web/requirements.txt
Django==1.9.7
gunicorn==19.3.0
psycopg2==2.6
redis==2.10.3
djangorestframework==3.3.3
